refactor: dnd holds the callbacks, not the work
The drag handlers had grown the low-level work inline: `dnd/pools.ts` alone
carried the preview node's SVG, the pointer clamping, the grid snapping, the
content sizing and the placement, so reading a callback meant reading all of
it. They now read as a few named steps, and the work sits where it can be
reached and named:

  utils/pools.ts       sizePoolToContent, isPoolBoundaryRequired,
                       getClampedPoolPosition, canMoveSwimlane,
                       findPoolViewAtPoint, getPositionInSwimlane
  effects/pool-preview show/move/removePoolPreview, beside ghost.ts and
                       swimlane-preview.ts, which already did this
  actions/place-pool   placeDroppedPool, dropPoolAt
  actions/insert-swimlane  insertSwimlaneIntoPool, dropSwimlaneIntoPool
  actions/add-element   addElementToSwimlane

`dnd/pools.ts` goes from 334 lines to 53, `dnd/swimlanes.ts` from 189 to
113, `dnd/elements.ts` from 247 to 102.

Names that were lying are fixed on the way: `checkSwimlaneDrop` performs the
drop rather than checking it, so it is `dropSwimlaneIntoPool`;
`positionInSwimlane` returns a point, so it is `getPositionInSwimlane`; and
`ensurePoolDragBoundary` is `getClampedPoolPosition`.

Two things the flattening turned up. `onSwimlaneDrag` derived the same
"is this pool acceptable" condition in three branches, once per early
return — it is computed once now. And `dropPoolAt` built its own first lane
instead of going through `insertSwimlaneIntoPool`, so "every pool gets a
lane" was written twice.
